# Tree-Diameter-Visualization/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type NodeLabel = number | string;

export interface TreeEdge {
  source: NodeLabel;
  target: NodeLabel;
}

export interface Tree {
  /**
     * @minimum 2
     * @maximum 1000
     */
  nodeCount: number;
  nodes: NodeLabel[];
  edges: TreeEdge[];
}

export interface TreeGenerateInput {
  /**
     * @minimum 2
     * @maximum 1000
     */
  nodes: number;
}

export type TreeInput = Tree;

export interface DiameterAnalysis {
  /** @minimum 1 */
  diameterLength: number;
  startNode: NodeLabel;
  endNode: NodeLabel;
  diameterPath: NodeLabel[];
  /**
     * @minimum 2
     * @maximum 1000
     */
  nodeCount: number;
  nodes: NodeLabel[];
  edges: TreeEdge[];
}