# Tree-Diameter-Visualization/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tree
    description: Tree generation and diameter analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tree/generate:
    post:
      operationId: generateTree
      tags: [tree]
      summary: Generate a valid random tree
      description: Generates a connected tree with the requested number of nodes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeGenerateInput"
      responses:
        "200":
          description: Generated tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tree"
        "400":
          description: Invalid node count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tree/diameter:
    post:
      operationId: calculateTreeDiameter
      tags: [tree]
      summary: Calculate a tree diameter
      description: Calculates the diameter path for a previously generated tree.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeInput"
      responses:
        "200":
          description: Diameter analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiameterAnalysis"
        "400":
          description: Invalid tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TreeEdge:
      type: object
      properties:
        source:
          $ref: "#/components/schemas/NodeLabel"
        target:
          $ref: "#/components/schemas/NodeLabel"
      required:
        - source
        - target
    Tree:
      type: object
      properties:
        nodeCount:
          type: integer
          minimum: 2
          maximum: 1000
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeLabel"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/TreeEdge"
      required:
        - nodeCount
        - nodes
        - edges
    TreeGenerateInput:
      type: object
      properties:
        nodes:
          type: integer
          minimum: 2
          maximum: 1000
      required:
        - nodes
    TreeInput:
      $ref: "#/components/schemas/Tree"
    DiameterAnalysis:
      type: object
      properties:
        diameterLength:
          type: integer
          minimum: 1
        startNode:
          $ref: "#/components/schemas/NodeLabel"
        endNode:
          $ref: "#/components/schemas/NodeLabel"
        diameterPath:
          type: array
          items:
            $ref: "#/components/schemas/NodeLabel"
        nodeCount:
          type: integer
          minimum: 2
          maximum: 1000
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeLabel"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/TreeEdge"
      required:
        - diameterLength
        - startNode
        - endNode
        - diameterPath
        - nodeCount
        - nodes
        - edges
    NodeLabel:
      oneOf:
        - type: integer
        - type: string
          pattern: '^-?(0|[1-9][0-9]*)$'