# Tree-Diameter-Visualization/lib/api-zod/src/generated/types/diameterAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NodeLabel } from './nodeLabel';
import type { TreeEdge } from './treeEdge';

export interface DiameterAnalysis {
  /** @minimum 1 */
  diameterLength: number;
  startNode: NodeLabel;
  endNode: NodeLabel;
  diameterPath: NodeLabel[];
  /**
     * @minimum 2
     * @maximum 1000
     */
  nodeCount: number;
  nodes: NodeLabel[];
  edges: TreeEdge[];
}