# Tree-Diameter-Visualization/lib/api-zod/src/generated/types/treeEdge.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NodeLabel } from './nodeLabel';

export interface TreeEdge {
  source: NodeLabel;
  target: NodeLabel;
}