# Tree-Diameter-Visualization/lib/api-zod/src/generated/types/treeGenerateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TreeGenerateInput {
  /**
     * @minimum 2
     * @maximum 1000
     */
  nodes: number;
}